refactor: Levellog 예외에 대한 구체 메세지 작성 및 `UnauthorizedException` common으로 패키지 변경
